+++ beast-of-argh/beast-of-argh.d/local/share/cook/bin/configure-consul.sh
@@ -27,17 +27,31 @@ sep=$'\001'
 #GOSSIPKEY="$(cat /mnt/consulcerts/gossip.key)"
 # GOSSIPKEY is passed in as a variable
 
-< "$TEMPLATEPATH/consul-agent.hcl.in" \
+# temporary removal HCL consul as not starting
+#< "$TEMPLATEPATH/consul-agent.hcl.in" \
+#  sed "s${sep}%%datacenter%%${sep}$DATACENTER${sep}g" | \
+#  sed "s${sep}%%nodename%%${sep}$NODENAME${sep}g" | \
+#  sed "s${sep}%%ip%%${sep}$IP${sep}g" | \
+#  sed "s${sep}%%consulservers%%${sep}$CONSULSERVERS${sep}g" \
+#  > /usr/local/etc/consul.d/agent.hcl
+#
+#chmod 600 \
+#  /usr/local/etc/consul.d/agent.hcl
+#echo "s${sep}%%gossipkey%%${sep}$GOSSIPKEY${sep}" | sed -i '' -f - \
+#  /usr/local/etc/consul.d/agent.hcl
+
+# temporary replacement with json config
+< "$TEMPLATEPATH/consul-agent.json.in" \
   sed "s${sep}%%datacenter%%${sep}$DATACENTER${sep}g" | \
   sed "s${sep}%%nodename%%${sep}$NODENAME${sep}g" | \
   sed "s${sep}%%ip%%${sep}$IP${sep}g" | \
   sed "s${sep}%%consulservers%%${sep}$CONSULSERVERS${sep}g" \
-  > /usr/local/etc/consul.d/agent.hcl
+  > /usr/local/etc/consul.d/agent.json
 
-chmod 600 \
-  /usr/local/etc/consul.d/agent.hcl
+chmod 600 /usr/local/etc/consul.d/agent.json
+
 echo "s${sep}%%gossipkey%%${sep}$GOSSIPKEY${sep}" | sed -i '' -f - \
-  /usr/local/etc/consul.d/agent.hcl
+  /usr/local/etc/consul.d/agent.json
 
 # set owner and perms on _directory_ /usr/local/etc/consul.d with agent.hcl
 chown -R consul:wheel /usr/local/etc/consul.d/
@@ -46,7 +60,8 @@ chown -R consul:wheel /usr/local/etc/consul.d/
 service consul enable || true
 
 # set load parameter for consul config
-sysrc consul_args="-config-file=/usr/local/etc/consul.d/agent.hcl"
+#sysrc consul_args="-config-file=/usr/local/etc/consul.d/agent.hcl"
+sysrc consul_args="-config-file=/usr/local/etc/consul.d/agent.json"
 sysrc consul_syslog_output_priority="warn"
 #sysrc consul_datadir="/var/db/consul"
 #sysrc consul_group="wheel"