+++ beast-of-argh/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Version bump for new base image
 * Update loki and promtail to 2.7.3
 * Set consul options for bind address
+* Revert to consul json config file until consul fixed start with hcl
 
 ---